changes in preparation for explore mode
