+++ src/components/LoadExplore/index.css
@@ -53,6 +53,7 @@
   flex-wrap: wrap;
   justify-content: stretch;
   align-items: stretch;
+  gap: 10px;
 }
 
 .row-input {
@@ -73,7 +74,7 @@
 }
 
 .text-100 {
-  color: #C87619;
+  /* color: #C87619; */
   font-weight: 500;
 }
